error: introduce BoxedError type
In later commits we will need the ability to box a generic error (on
pain of drowning in massive sets of generic constraints, blowing up
compile time and binary size).

Ideally we could just use a Box<dyn error::Error> but the Error trait is
not available on std
(tracking issue: https://github.com/rust-lang/rust/issues/103765 ).

However, when we *do* have std on, we'd like our error traits to work
naturally, not lose source information, etc. So we introduce a newtype
and have all its internals gated on std vs non-std, and expose internal
constructors that must be used on types that implement std::Error when
std is on, and implement core::fmt::Display when it's not.