expression: rewrite Tree module to no longer use a recursive data type
This significantly speeds up and simplifies tree parsing, at the cost of
having a more complicated API (but we mostly addressed the API question
in the previous commits).

This completely eliminates recursion for the Tree data type, including
in the Drop impl.

Big diff but there are only two "real" changes -- expression/mod.rs is
substantially rewritten of course since we replace the core datatype,
and Tr::from_tree is substantially rewritten since doing so was the
point of this change. The rest of the changes are mechanically changing
the signature of expression::FromTree::from_tree everywhere.