miniscript: introduce ValidationParams type
This introduces the ValidationParams type but does not use it anywhere.
As you can see, this is a composite of the checks in ExtParams and the
checks in miniscript::analyzable and the checks in miniscript::context.

We add several associated constants to each context. The eventual goal
is that we remove the existing analyzable/ExtParams/Ctx stuff, drop Ctx
as a type parameter on the Miniscript type, and then the Ctx trait
becomes basically a nice way to access constants.

I am keeping the terms sane/insane throughout these PRs. Once all the
dust has settled, these names will appear in the names of parse/decode
methods and in the names of constants in Ctx, and they should be used
consistently. At that point we can revisit whether we want to rename
them and/or add more. (I propose that "sane" become "default", that
"insane" become "consensus" and allow raw pkhs, and that we add some
standardness ones that maybe have dates/versions on them in case the
most popular standardness rules change in future. But we can leave that
for later.)

A natural question when seeing this structure is "why not just use
ExtParams". Well, I could have, but I hate the name of the struct and
the names of all the fields. And since I'm causing a lot of API breakage
I might as well take the opportunity to improve the naming.