miniscript: replace from_str_ext with from_str_with_validation_params
This moves us toward being able to drop ExtParams. It also eliminates
the NonTopLevel error variant, and replaces the "hack" in Tr::from_tree
with a more sensible "do all the sanity checks". It is still
wrong/inconsistent that in Tr::from_tree we are doing validation while
in every other from_tree instance we need to validate afterward, but we
are moving in the right direction. A later commit will change the
FromTree trait to incorporate validation.