remove all sanity_check methods from the library
Well, there are sanity checks in the type system (which are private and
make sense) and a sanity_check method in the PSBT module that I didn't
look at. But the rest are gone now.

These should really never have existed. At this point in this commit
series we are pretty close to a point where they serve no purpose at
all; miniscripts constructed with from_ast, from_str and decode are
all sanity-checked at construction with the same set of checks.

We haven't yet attempted to apply these validation parameters to either
of the policy types, or gone through all the non-miniscript descriptor
types, but we'll get there.