miniscript: delete ExtParams, Ctx::check_* methods, and error types
This is a big commit but it's almost all red.

This results in a bunch of error messages changing since we now fail due
to ValidationParams rather than due to Context::check_*.

One observable performance change is that we parse entire Miniscripts
before doing bounds checks. (This means that the error messages for
exceeded limits are now more accurate, though they will probably revert
when we do some optimization passes.)

There is also a performance hit because we are now doing validity checks
in the compiler: in insert_elem we replace a call to
Ctx::check_local_validity with a call to ms.validate_non_top_level, and
this causes a duplicate pubkey check (and some other expensive checks)
to be run even though these checks pass by construction. We will come
back to this and fix the performance!

There is a max recursion check in Expression::from_str and another one
in Miniscript::from_ast that I have left alone, for now, because of
this. The recursion check needs to fire immediately before constructing
any object that exceeds the maximum depth.

Later we will add a call to validate_non_top_level to Miniscript::from_ast.
This will let us drop the maximum depth check, but it will require an API
change because the caller will need to provide validation parameters.
So we leave it be for now.