expression: change from_str to return a ParseError
The goal is that when parsing an expression into a miniscript (or
whatever) that there are only two kinds of errors: validation and
parsing. To get there, we need to stop returning Error from expression
parsing methods.

from_str arguably should return a ParseTreeError, which is a specific type
inside the expression module, but for now we stick with ParseError to
minimize the (already large) number of .map_err calls we need to add.

This required changing some expression unit tests to remove one layer of
error wrapper types. While I was in there I uncommented a couple tests
that I'd meant to uncomment as part of the expression overhaul PRs
several months ago. They pass :).