miniscript: introduce ConstructError to return from from_ast
In order to clean up the error types of Miniscript::decode and the
various from_str methods, we need to clean up the error return of
from_ast, which is the primary generic constructor of Miniscript nodes.

It can fail either by a typecheck error or a recursive depth check
error. Going forward, the recursive depth check error will be folded
into a validation check and the user will be able to specify which
parameters to use.

To stop some unit tests from breaking, we allow construction of raw
pkhs for now.

Eliminates the MaxRecursiveDepthExceeded error from the top-level error
enum. Doing this made me notice that our unit test for this wasn't doing
anything! It calls `matches!` instead of `assert!(matches!)`. Derp.