miniscript: use ValidationError for key translation
The "outer error" we yield when translating keys is currently the
top-level Error type. This is much too general. In fact the only
possible errors that can occur are validation errors (which, thanks to
these last PRs, encompass everything from incorrect keys for context,
to duplicate keys, to exceeding script limits).

In the past it has been difficult to make this change because the
implementation of Miniscript::translate_pk_ctx internally calls
Miniscript::from_ast to reconstruct the script. This method would return
the top-level Error type and it was difficult to convince myself that it
would only return validation errors and never anything else.

With the new ConstructError type, there are only two possibilities:
validation errors and typecheck errors. And it is (somewhat) easy to
convince yourself that typecheck errors cannot ever happen due to key
translation, since the only typecheck errors are about script opcode
semantics, which aren't affected by any properties of the keys.