expression: delete FromTree trait
This doesn't need to be a trait. We don't use it generically and it's
hard to imagine where we would. If there is demand for it we probably
want to use TryFrom instead.

Change all the from_tree methods to inherent methods. This makes them
easier to call, and more importantly will let me change the error return
on some without needing to change them all at once.