miniscript: cache validation params in miniscript
Ok, we cache which checks we've done, but actually in this commit we don't
actually cache anything -- we always set the set of already-validated
parameters to "MAX", i.e. no checks. But we add the logic to do this, and
in the next commit we will modify all our constructors to include validation
params.