miniscript: add multipath len to types::ExtData, check it directly
This means we can check for multipath lengths non-recursively (in fact,
in constant time).

In the next commit we are going to start caching which validation parameters
have already been checked, in order to avoid n^2 work in the common case where
the user builds up a Miniscript node by node, validating each one as they go.
The idea is that if both children of a node have been checked for, e.g. no
x-only public keys, then we immediately know that the node doesn't have any,
without needing to iterate over it checking every key.

However, the "no inconsistent multipath lengths" check can't be cached in this
way, since even if the two children of a node individually have consistent
multipath lengths, it may be that they're inconsistent with each other. So we
need a ext type parameter to capture this.

BTW, the name "multipath length" is pretty bad. I guess we should rename it.
Not sure what would be better. "Width" maybe?