miniscript: change types::ExtParams::multi_a to take threshold
We will need the pubkey data in the next commit in order to count multipath
lengths. This function previously took k and n as usizes, probably because
I was lazy and didn't want to type out a Threshold<N, Pk> signature, but my
laziness has come back to bite me.