errors: introduce WithSpan type, use it for Miniscript type errors
We currently have only one error type (miniscript::types::Error that
has any span information. This is implemented in an ad-hoc way. We
want to allow attaching span information to any ConstructError. To
do this, we introduce a generic WithSpan type.

(We also have span information in expression::ParseTreeError but that
is a much more complex scenario and isn't just "error plus a span").

Despite the name, `WithSpan` actually carries individual positions.
But it felt more discoverable to put "span" in the name since people
associated spans with error messages.