miniscript: bump span error information up to ConstructError
Rather than only having type errors have span info, also add it to
validation errors (especially since validation errors are much more
likely to require thought for users to understand).

For now we just stick the string serialization of the Minischipt
under construction into the error. When parsing from strings, we
should put the whole string into the error and mark the position
of where the actual problem was.

In followups we may want to try to add some clever truncation
logic or other display improvements.