put PartialEq and Eq on all errors in the library
This deepens the "compromises we have to make because Rust's error
types are a mass" situation but I added a big block comment and I
think that in practice this is totally fine and nobody will ever
notice.

Meanwhile, this lets us use `assert_eq` on error types rather than
`assert!(matches!`, which in turn will allow us to define rich error
types that aren't just enums you can pattern-match on.