sortedmulti: deprecate sorted_node in favor of sorted_threshold
The sorted_node function basically just wraps the threshold in a Terminal,
which you then can't do anything except convert to a Miniscript using
from_ast, which returns a Result which can't error.

It's much simpler to just return the threshold directly and let the
user decide what to do with it.