f miniscript: allow `decode`ing arbitrary pubkeys
Ok, not arbitrary pubkeys, but any that implements the ParseableKey trait,
which was previously a trait that gave us a `from_slice` method, but now
gives us three methods for decoding from various array sizes. This lets
us directly capture "is this an x-only key, from the perspective of the
decoder" in the trait.

We also unseal the trait, since I don't see any reason we need to restrict
it to just bitcoin::PublicKey and secp256k1::XOnlyPublicKey anymore.

f have a couple fixmes