interpreter: stop re-encoding miniscripts that were decoded from slices
In the next commit we will remove the dependence on Ctx from the decoder
(and interpreter). To do this while ensuring that we only accept the
correct kinds of keys from the interpreter, we need to decode to a
Miniscript which uses the interpreter::BitcoinKey type.

But interpreter::BitcoinKey does not impl ToPublicKey or ParseableKey.

We can make it implement ParseableKey by fixing that trait to more
accurately reflect what's needed in Miniscript::decode; we can then
unseal the trait, drop the ToPublicKey requirement, and relax the
decoder logic to allow decoding arbitrary key/context combinations.

But then we lose the ability to re-encode decoded scripts. Which is
fine, because we don't need to do this ... after refactoring a bit.
Which is what this commit does.