miniscript: stop using Ctx::pk_len in miniscript/descriptor
Replace Ctx::pk_len with a method on ValidationParams. We then need to
add a &ValidationParams argument to Miniscript::maximum_script_size (and
SortedMulti::maximum_script_size which is an annoying copy of this).

As will be a pattern in the next couple commits, we see that:

* In generic contexts we have a `params` object available to use
* In tests and in the descriptor module, we are in a specific context
  and we can just hardcode e.g. Segwitv0::SANE.

There is a tension highlighted by this commit: ValidationParams allows
you to set all key types (or no key types) to 'allowed' simultaneously,
even though no on-chain environment allows both x-only and non-x-only
keys.

In this case, we are estimating sizes so we can just assume the worst,
which is fine, but over the next several commits we will replace other
instances of SigType where we have to decide "are signatures ECDSA or
are they Schnorr?" and it's unclear how we ought to behave if both
are allowed.

The tension is that ValidationParams is designed to be monotonic, so
you can always do params1.intersect(params2) and get a stricter set
of validation parameters. This design goal prevents us from encoding
"either-or" conditions like SigType, which in turn forces some parts
of our logic to think about the no-keys/all-keys situation even
though that is absurd.