miniscript: use ValidationParams rather than Ctx::sig_type in pk_k
We change the method ExtData::pk_k to use ValidationParams, which causes
API changes that bubble outward. As you can see, in most cases we have
params available to use.

In the next commit we'll make the same change for pk_h, which will be a
much smaller change since we did the "hard work" here, and will eliminate
the Ctx parameter from the type checking module entirely.

Again, in this commit we're just doing worst-case estimation so there
isn't a problem with "what if both x-only and full keys are allowed"?