typeck: stop using Ctx in ExtData::pk_h
We are no longer using the Ctx object at all in the Miniscript type checker.
(You can 'git grep' in src/miniscript/types and you will see that Ctx only
appears as part of the type signature for Terminal.)

The next commits will continue to eliminate Ctx::sig_type where it is called.