miniscript: change tap_leaf_internal to return an option
We now return an Option<TapLeafHash> from tap_leaf_internal, and use the
optionality of this to determine whether we are doing a Schnorr signature
or an ECDSA signature. This lets us remove a Ctx parameter from the
internal Witness::signature method, and several methods that call it.
Also eliminates some logic which theoretically could lead to an ECDSA
signature produced for multi_a or a Schnorr sig for multi; now we have
one new `expect()` which unwraps the Option for multi_a.

There are still Ctx parameters in Witness::pkh_signature and pkh_public_key,
which are used for key size estimation. These will be eliminate when we
introduce ValidationParams, which will do size estimation in a more flexible
way.